Separation core/cli
