+++ ghga_connector/core/__init__.py
@@ -18,32 +18,6 @@ This sub-package contains the main business functionality of this service.
 It should not contain any service API-related code.
 """
 
-from .api_calls import (  # noqa: F401
-    UploadStatus,
-    await_download_url,
-    download_api_call,
-    get_file_metadata,
-    get_part_upload_urls,
-    initiate_multipart_upload,
-    patch_multipart_upload,
-    start_multipart_upload,
-)
-from .exceptions import (  # noqa: F401
-    BadResponseCodeError,
-    CantChangeUploadStatus,
-    KnownError,
-    MaxRetriesReached,
-    MaxWaitTimeExceeded,
-    NoS3AccessMethod,
-    NoUploadPossibleError,
-    RequestFailedError,
-    UploadNotRegisteredError,
-    UserHasNoUploadAccess,
-)
-from .file_operations import (  # noqa: F401
-    download_file_parts,
-    read_file_parts,
-    upload_file_part,
-)
-from .main import check_url  # noqa: F401
+
+from .main import MAX_RETRIES, download_core, upload_core  # noqa: F401
 from .message_display import AbstractMessageDisplay, MessageColors  # noqa: F401
